+++ ydb/core/kqp/executer_actor/kqp_streaming_helper.cpp
@@ -1,0 +1,51 @@
+#include "kqp_streaming_helper.h"
+
+namespace NKikimr::NKqp {
+
+namespace {
+
+struct TStreamingQueryCounters : public IStreamingQueryCounters {
+
+    TStreamingQueryCounters(const ::NMonitoring::TDynamicCounterPtr& counters, const TString& path)
+        : Path(path)
+    {
+        auto subGroup = counters->GetSubgroup("subsystem", "streaming_queries");
+        SubGroup = subGroup->GetSubgroup("path", Path);
+        CpuMs = SubGroup->GetCounter("streaming.query.cpu.usage.milliseconds");
+        MemoryUsageBytes = SubGroup->GetCounter("streaming.query.memory.usage.bytes");
+        UptimeSeconds = SubGroup->GetCounter("streaming.query.uptime.seconds");
+        TaskCount = SubGroup->GetCounter("streaming.query.tasks.count");
+        InputBytes = SubGroup->GetCounter("streaming.query.input.bytes");
+        OutputBytes = SubGroup->GetCounter("streaming.query.output.bytes");
+    }
+
+    ~TStreamingQueryCounters() {
+        SubGroup->RemoveSubgroup("path", Path);
+    }
+
+    void Update(const TAggExecStat& stats) {
+        CpuMs->Set(stats.CpuTimeMs);
+        MemoryUsageBytes->Set(stats.MaxMemoryUsageBytes);
+        UptimeSeconds->Set(stats.DurationSeconds);
+        TaskCount->Set(stats.TasksCount);
+        InputBytes->Set(stats.InputBytes);
+        OutputBytes->Set(stats.OutputBytes);
+    }
+
+    const TString Path;
+    ::NMonitoring::TDynamicCounterPtr SubGroup;
+    ::NMonitoring::TDynamicCounters::TCounterPtr CpuMs;
+    ::NMonitoring::TDynamicCounters::TCounterPtr MemoryUsageBytes;
+    ::NMonitoring::TDynamicCounters::TCounterPtr UptimeSeconds;
+    ::NMonitoring::TDynamicCounters::TCounterPtr TaskCount;
+    ::NMonitoring::TDynamicCounters::TCounterPtr InputBytes;
+    ::NMonitoring::TDynamicCounters::TCounterPtr OutputBytes;
+};
+
+} // namespace
+
+TIntrusivePtr<IStreamingQueryCounters> MakeStreamingQueryCounters(const ::NMonitoring::TDynamicCounterPtr& counters, const TString& path) {
+    return MakeIntrusive<TStreamingQueryCounters>(counters, path);
+}
+
+} // namespace NKikimr::NKqp
